// src/db/schema.ts
export interface Board {
  id: string;
  title: string;
  subtitle: string | null;
  description: string | null;
  nsfw: number;
  max_threads: number;
  bump_limit: number;
  created_at: number;
}

export interface Thread {
  id: number;
  board_id: string;
  subject: string;
  sticky: number;
  locked: number;
  post_count: number;
  image_count: number;
  created_at: number;
  bumped_at: number;
}

export interface Post {
  id: number;
  thread_id: number;
  board_id: string;
  is_op: number;
  author_name: string;
  tripcode: string | null;
  poster_hash: string | null;
  comment: string | null;
  file_url: string | null;
  file_name: string | null;
  file_size: number | null;
  file_width: number | null;
  file_height: number | null;
  file_type: string | null;
  ip_hash: string | null;
  sage: number;
  created_at: number;
}

export interface Ban {
  id: number;
  ip_hash: string;
  reason: string | null;
  board_id: string | null;
  expires_at: number;
  created_at: number;
}

export interface ModLog {
  id: number;
  action: string;
  target_type: string | null;
  target_id: string | null;
  details: string | null;
  created_at: number;
}

export interface ThreadWithOP extends Thread {
  op_post?: Post;
  recent_posts?: Post[];
}

export interface CatalogItem extends Thread {
  op_post?: Post;
}